use REGISTER_OP macro
