+++ chainerx_cc/chainerx/cuda/cuda_device/activation.cu
@@ -123,25 +123,5 @@ void CudaDevice::Tanh(const Array& x, const Array& out) {
     });
 }
 
-namespace {
-
-template <typename T>
-struct FloorImpl {
-    using CudaType = cuda_internal::DataType<T>;
-    __device__ void operator()(int64_t /*i*/, CudaType x, CudaType& out) { out = cuda::Floor(x); }
-};
-
-}  // namespace
-
-void CudaDevice::Floor(const Array& x, const Array& out) {
-    CheckDevicesCompatible(x, out);
-    CudaSetDeviceScope scope{index()};
-    const Array& x_cast = x.dtype() == out.dtype() ? x : x.AsType(out.dtype());
-    VisitFloatingPointDtype(out.dtype(), [&](auto pt) {
-        using T = typename decltype(pt)::type;
-        Elementwise<const T, T>(FloorImpl<T>{}, x_cast, out);
-    });
-}
-
 }  // namespace cuda
 }  // namespace chainerx
