+++ chainerx_cc/chainerx/cuda/cuda_device/misc.cu
@@ -10,8 +10,10 @@
 #include "chainerx/cuda/cuda_set_device_scope.h"
 #include "chainerx/cuda/elementwise.cuh"
 #include "chainerx/cuda/numeric.cuh"
+#include "chainerx/cuda/op_regist.h"
 #include "chainerx/device.h"
 #include "chainerx/dtype.h"
+#include "chainerx/routines/math.h"
 
 namespace chainerx {
 namespace cuda {
@@ -91,6 +93,28 @@ void CudaDevice::IsInf(const Array& x, const Array& out) {
         Elementwise<const T, bool>(IsInfImpl<T>{}, x, out);
     });
 }
+
+template <typename T>
+struct FloorImpl {
+    using CudaType = cuda_internal::DataType<T>;
+    __device__ void operator()(int64_t /*i*/, CudaType x, CudaType& out) { out = cuda::Floor(x); }
+};
+
+class CudaFloorOp : public FloorOp {
+protected:
+    void Impl(const Array& x, const Array& out) override {
+        Device& device = x.device();
+        device.CheckDevicesCompatible(x, out);
+        CudaSetDeviceScope scope{device.index()};
+        const Array& x_cast = x.dtype() == out.dtype() ? x : x.AsType(out.dtype());
+        VisitFloatingPointDtype(out.dtype(), [&](auto pt) {
+            using T = typename decltype(pt)::type;
+            Elementwise<const T, T>(FloorImpl<T>{}, x_cast, out);
+        });
+    }
+};
+
+CHAINERX_REGISTER_OP_CUDA(FloorOp, CudaFloorOp);
 
 }  // namespace cuda
 }  // namespace chainerx
